Docker-DCO-1.1-Signed-off-by: Dan Walsh <dwalsh@redhat.com> (github: rhatdan)

+++ pkg/archive/archive_test.go
@@ -873,7 +873,8 @@ func getNlink(path string) (uint64, error) {
 	if !ok {
 		return 0, fmt.Errorf("expected type *syscall.Stat_t, got %t", stat.Sys())
 	}
-	return statT.Nlink, nil
+	// We need this conversion on ARM64
+	return uint64(statT.Nlink), nil
 }
 
 func getInode(path string) (uint64, error) {
